requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy==1.4.51
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==1.10.7
